@@ -1,101 +1,121 @@
+import path from 'path';
+
 import globby from 'globby';
 import pLimit from 'p-limit';
-import isGlob from 'is-glob';
-import path from 'path';
 import minimatch from 'minimatch';
-import writeFile from './writeFile';
+
 import isObject from './utils/isObject';
 
 export default function processPattern(globalRef, pattern) {
-    const {info, debug, output, concurrency, contextDependencies} = globalRef;
-    const globArgs = Object.assign({
-        cwd: pattern.context
-    }, pattern.fromArgs || {});
+  const { logger, output, concurrency, compilation } = globalRef;
+  const globOptions = Object.assign(
+    {
+      cwd: pattern.context,
+      follow: true,
+      // Todo in next major release
+      // dot: false
+    },
+    pattern.globOptions || {}
+  );
 
-    if (pattern.fromType === 'nonexistent') {
-        return Promise.resolve();
-    }
+  if (pattern.fromType === 'nonexistent') {
+    return Promise.resolve();
+  }
 
-    const limit = pLimit(concurrency || 100);
+  const limit = pLimit(concurrency || 100);
 
-    info(`begin globbing '${pattern.glob}' with a context of '${pattern.context}'`);
-    return globby(pattern.glob, globArgs)
-        .then((paths) => Promise.all(paths.map((from) => limit(() => {
-            const file = {
-                force: pattern.force,
-                absoluteFrom: path.resolve(pattern.context, from)
+  logger.info(
+    `begin globbing '${pattern.glob}' with a context of '${pattern.context}'`
+  );
+
+  return globby(pattern.glob, globOptions).then((paths) =>
+    Promise.all(
+      paths.map((from) =>
+        limit(() => {
+          const file = {
+            force: pattern.force,
+            absoluteFrom: path.resolve(pattern.context, from),
+          };
+
+          file.relativeFrom = path.relative(pattern.context, file.absoluteFrom);
+
+          if (pattern.flatten) {
+            file.relativeFrom = path.basename(file.relativeFrom);
+          }
+
+          logger.debug(`found ${from}`);
+
+          // Check the ignore list
+          let il = pattern.ignore.length;
+
+          // eslint-disable-next-line no-plusplus
+          while (il--) {
+            const ignoreGlob = pattern.ignore[il];
+
+            let globParams = {
+              dot: true,
+              matchBase: true,
             };
-            file.relativeFrom = path.relative(pattern.context, file.absoluteFrom);
 
-            if (pattern.flatten) {
-                file.relativeFrom = path.basename(file.relativeFrom);
+            let glob;
+
+            if (typeof ignoreGlob === 'string') {
+              glob = ignoreGlob;
+            } else if (isObject(ignoreGlob)) {
+              glob = ignoreGlob.glob || '';
+
+              const ignoreGlobParams = Object.assign({}, ignoreGlob);
+              delete ignoreGlobParams.glob;
+
+              // Overwrite minimatch defaults
+              globParams = Object.assign(globParams, ignoreGlobParams);
+            } else {
+              glob = '';
             }
 
-            // This is so webpack is able to watch the directory and when
-            // a new file is added it triggeres a rebuild
-            const contextPath = path.dirname(path.resolve(from));
-            if (contextDependencies.indexOf(contextPath) === -1 && isGlob(pattern.glob)) {
-                contextDependencies.push(contextPath);
+            logger.debug(`testing ${glob} against ${file.relativeFrom}`);
+
+            if (minimatch(file.relativeFrom, glob, globParams)) {
+              logger.info(
+                `ignoring '${file.relativeFrom}', because it matches the ignore glob '${glob}'`
+              );
+
+              return Promise.resolve();
             }
 
-            debug(`found ${from}`);
+            logger.debug(`${glob} doesn't match ${file.relativeFrom}`);
+          }
 
-            // Check the ignore list
-            let il = pattern.ignore.length;
-            while (il--) {
-                const ignoreGlob = pattern.ignore[il];
+          // Change the to path to be relative for webpack
+          if (pattern.toType === 'dir') {
+            file.webpackTo = path.join(pattern.to, file.relativeFrom);
+          } else if (pattern.toType === 'file') {
+            file.webpackTo = pattern.to || file.relativeFrom;
+          } else if (pattern.toType === 'template') {
+            file.webpackTo = pattern.to;
+            file.webpackToRegExp = pattern.test;
+          }
 
-                let globParams = {
-                    dot: true,
-                    matchBase: true
-                };
+          if (path.isAbsolute(file.webpackTo)) {
+            if (output === '/') {
+              const message =
+                'using older versions of webpack-dev-server, devServer.outputPath must be defined to write to absolute paths';
 
-                let glob;
-                if (typeof ignoreGlob === 'string') {
-                    glob = ignoreGlob;
-                } else if (isObject(ignoreGlob)) {
-                    glob = ignoreGlob.glob || '';
-                    const ignoreGlobParams = Object.assign({}, ignoreGlob);
-                    delete ignoreGlobParams.glob;
+              logger.error(message);
 
-                    // Overwrite minimatch defaults
-                    globParams = Object.assign(globParams, ignoreGlobParams);
-                } else {
-                    glob = '';
-                }
-
-                debug(`testing ${glob} against ${file.relativeFrom}`);
-                if (minimatch(file.relativeFrom, glob, globParams)) {
-                    info(`ignoring '${file.relativeFrom}', because it matches the ignore glob '${glob}'`);
-                    return Promise.resolve();
-                } else {
-                    debug(`${glob} doesn't match ${file.relativeFrom}`);
-                }
+              compilation.errors.push(new Error(message));
             }
 
-            // Change the to path to be relative for webpack
-            if (pattern.toType === 'dir') {
-                file.webpackTo = path.join(pattern.to, file.relativeFrom);
-            } else if (pattern.toType === 'file') {
-                file.webpackTo = pattern.to || file.relativeFrom;
-            } else if (pattern.toType === 'template') {
-                file.webpackTo = pattern.to;
-                file.webpackToRegExp = pattern.test;
-            }
+            file.webpackTo = path.relative(output, file.webpackTo);
+          }
 
-            if (path.isAbsolute(file.webpackTo)) {
-                if (output === '/') {
-                    throw '[copy-webpack-plugin] Using older versions of webpack-dev-server, devServer.outputPath must be defined to write to absolute paths';
-                }
+          logger.info(
+            `determined that '${from}' should write to '${file.webpackTo}'`
+          );
 
-                file.webpackTo = path.relative(output, file.webpackTo);
-            }
-
-            // ensure forward slashes
-            file.webpackTo = file.webpackTo.replace(/\\/g, '/');
-
-            info(`determined that '${from}' should write to '${file.webpackTo}'`);
-
-            return writeFile(globalRef, pattern, file);
-        }))));
+          return file;
+        })
+      )
+    )
+  );
 }
